experiments/misc/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
black==23.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
dataclasses-json==0.5.7
datasets==2.9.0
dill==0.3.6
et-xmlfile==1.1.0
faiss-cpu==1.7.3
filelock==3.9.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
huggingface-hub==0.12.0
idna==3.4
InstructorEmbedding==1.0.0
joblib==1.2.0
jsonlines==3.1.0
kiwisolver==1.4.4
langchain==0.0.77
markdown-it-py==2.1.0
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.6.3
mdurl==0.1.2
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==0.4.3
nltk==3.8.1
numpy==1.24.1
openai==0.25.0
openpyxl==3.1.0
packaging==23.0
pandas==1.5.3
pandas-stubs==1.5.3.230203
pathspec==0.11.0
Pillow==9.4.0
platformdirs==2.6.2
plotly==5.13.0
pyarrow==11.0.0
pydantic==1.10.4
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
responses==0.18.0
rich==13.3.1
scikit-learn==1.2.1
scipy==1.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
SQLAlchemy==1.4.46
tenacity==8.1.0
threadpoolctl==3.1.0
tokenizers==0.12.1
tomli==2.0.1
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
transformers==4.20.0
types-pytz==2022.7.1.0
typing-inspect==0.8.0
typing_extensions==4.4.0
urllib3==1.26.14
xxhash==3.2.0
yarl==1.8.2
